Update HOLMAN_TAP_TIMER.md
Fix schedule switches to ensure correctness. Not sure if the DPS have changed slightly since this was written, but the previous schedule schedule switch would cause weird behaviour.  It would incorrectly manipulate the data, corrupting the encoded value (so when the switch is toggled back, the original schedule is lost). This also caused index errors.